Fix some bugs in the exponential backoff.
git-svn-id: svn+ssh://spartan.csl.sri.com/svn/private/DEFIANCE@216 a58ff0ac-194c-e011-a152-003048836090

+++ src/rng.cc
@@ -216,7 +216,7 @@ rng_range_geom(unsigned int hi, unsigned int xv)
      ( e^{-hi/xe}, 1 ]. Doing this with arithmetic introduces
      a slight nonuniformity, but we really want to avoid rejection
      sampling here. */
-  double ulo = exp(-hi/xe);
+  double ulo = exp(-double(hi)/xe);
   U = ulo + U * (1-ulo);
 
   /* Inverse transform sampling gives us a value for the exponential
